admin delete

+++ placeholder-frontend/src/services/UserService.js
@@ -69,14 +69,25 @@ export const updateUser = ({ user, values, profilePic, coverUrl }) => {
   });
 };
 
-export function deleteUser(user_id, cs_mail, user_password){
-  return client.delete ( "/deleteUser", {
+export function deleteUser(user_id, cs_mail, user_password, user_type) {
+  if (parseInt(user_type) === 0) {
+    return client.delete("/deleteUser", {
+      data: {
+        user_id: user_id,
+        cs_mail: cs_mail,
+        user_password: user_password,
+        user_type: user_type,
+      },
+    });
+  }
+  return client.delete("/deleteUser", {
     data: {
-      "user_id" : user_id,
-      "cs_mail" : cs_mail,
-      "user_password" : user_password,
+      user_id: user_id,
+      cs_mail: cs_mail,
+      user_password: user_password,
+      user_type: "-1",
     },
-  })
+  });
 }
 
 export function getUsersConnected(user_id) {
